monitoring/requirements.txt
flask==2.0.1
flask-cors==3.0.10
websockets==10.1
psycopg2-binary==2.9.1
prometheus-client==0.11.0
werkzeug==2.0.1
psutil==5.9.5